requirements.txt
aiofiles==0.7.0
viaa-chassis==0.2.0rc1
fastapi==0.63.0
fastapi-route-logger-middleware==0.1.3
uvicorn==0.13.4
requests==2.25.1
ldap3==2.9
zenpy>=2.0
slack-sdk==3.8.0
redis==3.5.3
apscheduler==3.7.0
